Allowed gen_diffs to work better.
[Details]
Added an option to gen_diffs.sh to stop it from capping the size of the
diff rename checker. This allows gen_diffs to better detect renames.
Also gave the option for gen_diffs to be called without arguments, and
generate the full git history instead (more stable, less efficient).

[File's changed]

	modified:   gen_diffs.sh