Sped up the backtracker.
[Details]
Changed the skip method to use StringBuilder instead of Strings to store
what is skipped.

[Note]
A large amount of time is spent in the skip function skipping over
content that is not used. Unfortunately, this method is also reused for
skipping over and returning content that will be used in other parts of
the backtracking algorithm. Strings were therefore used to save this
content, however the useless content being skipped over would spend long
time periods creating these strings. Using StringBuilder to create these
strings makes this process enormously more efficient, and thus capable
of processing even the unnecessary content.

[Bugs]

[File's changed]

	modified:   DiffSet.java