Added support for combined diffs.
[Details]
Added the support for the combined diff format, making sure and only
accepting this format when the code in this format is new to all parents
(i.e. added in the merge commit).

[Note]
Previously the diff generator generated diffs by diffing two consecutive
versions. This is sufficient if the commit history is linear, but
branching will confuse this strategy. Instead, we now produce commit
diffs and use these for backtracking. However doind so sometimes leads to
combined diffs for merge commits, which include changes introduced from
merge conflicts. To solve this, support was added for combined diffs to
the backtracker.

[File's changed]

	modified:   Diff.java