Changed to using forward diffs.
[Details]
To to the nature of merge commits (which were not initially considered),
it is necessary now to switch to using the forward diffs in the
backtracking algorithm instead of the backwards ones. This is a somewhat
minor change to the implementation, however.

[Bugs]
There seems to be some bugs still lingering with these new chnages. Bug
fixes to follow in following commits.

[File's changed]

	modified:   ../backtrack.jar
	modified:   Diff.java
	modified:   DiffSet.java
	modified:   ../gen_diffs.sh