Fixed minor bug.
[Details]
Fixed a small bug introduced by the change to using forward diffs.

[Bugs]
The bug was due to forward diffs having the adding of a file now
representing that file being unavailable in the history before that, and
the backtracking code did not account for a difference in the commit
diffs structure properly for this.

[File's changed]

	modified:   backtrack.jar
	modified:   backtrack/DiffSet.java