Changed backtracker to respect full git history.
[Details]
Complete overhaul of the backtracking algorithm that now takes the full
git history and traverses the history as a graph instead of a linear
history. This theoretically will allow the backtracker to always get the
backtracking right as it considers the full git tree when backtracking
the lines.

[Note]
In order for the backtracker to respect full git history, the shas file
given to the backtracker is now updated to have the sha of each commit,
the commits parents, and the ancestor of the commits parents (if
applicable). This allows the backtracker to construct a full git history
tree.

[Bugs]
There are a few major bugs in this version. Particularly, it seems that
despite only intending to visit a node once (not in psuedo), the
backtracker is visiting certain nodes much more. This bug could stem
from the complex git histories in the BugsInPy projects that trigger
incorrect functionality.

[File's changed]

	modified:   backtrack.jar
	modified:   backtrack/Backtrack.java
	modified:   backtrack/BugFile.java
	modified:   backtrack/BugFix.java
	new file:   backtrack/Node.java
	modified:   gen_diffs.sh