Manually updated backtracking results.
[Details]
Manually edited the backtracking results to all be correct. This breaks
the automation of the backtracking tool, however it gives consistently
correct results for all the BugsInPy projects without having to model
the complex git history trees (see the full_history branch for that
feature).

[Note]
Still to check the Defects4J projects.

[File's changed]

	modified:   cookiecutter_backtrack.json
	modified:   fastapi_backtrack.json
	modified:   keras_backtrack.json
	modified:   luigi_backtrack.json
	modified:   matplotlib_backtrack.json
	modified:   sanic_backtrack.json
	modified:   scrapy_backtrack.json
	modified:   thefuck_backtrack.json
	modified:   tornado_backtrack.json
	modified:   tqdm_backtrack.json
	modified:   tqdm_src.txt
	modified:   youtube-dl_backtrack.json
	new file:   ../checks.txt