Fixed the backtracking algorithm for start diffs.
[Details]
The Defects4J dataset presented a problem with backtracking based on
the differences between the actual buggy version and Defects4J buggy
version. This was due to the diff between these two usually touching the
buggy lines, as these lines were made to look more "buggy" by the
Defects4J developers. Because of this, these lines should have been
removed, but the backtracking algorithm had been set not to remove lines
from start diffs. To remedy this, an LCS algorithm is now being used to
find the best matching line from the actual buggy version that matches
the buggy lines in the Defects4J buggy version. This algorithm is
currently only set to be used when dealing with start diffs, however it
could easily be implemented as a command line parameter to allow this to
be used for the entire backtracking, increasing the likelihood of
backtracks.

[File's changed]
	modified:   ../backtrack.jar
	modified:   BugFile.java
	modified:   Chunk.java
	modified:   Diff.java
	new file:   LCS.java
	modified:   SearchList.java